add role demo improvement and remove display demo
