+++ docs/storybook/stories/Demo/DisplayColorTesting.stories.tsx
@@ -64,48 +64,3 @@ export const DisplayDemo = () => {
     </Box>
   )
 }
-
-const roles = [
-  'neutral',
-  'accent',
-  'success',
-  'open',
-  'danger',
-  'closed',
-  'attention',
-  'severe',
-  'done',
-  'upsell',
-  'sponsors',
-]
-
-export const RoleDemo = () => {
-  return (
-    <Box sx={{gap: 3, display: 'flex', flexDirection: 'column'}}>
-      <Box sx={{gap: 2, display: 'flex'}}>
-        {roles.map(role => (
-          <Token
-            text={role}
-            sx={{
-              backgroundColor: `var(--bgColor-${role}-muted)`,
-              color: `var(--fgColor-${role})`,
-              borderColor: `var(--borderColor-${role}-emphasis)`,
-            }}
-          />
-        ))}
-      </Box>
-      <Box sx={{gap: 2, display: 'flex'}}>
-        {roles.map(role => (
-          <Token
-            text={role}
-            sx={{
-              backgroundColor: `var(--bgColor-${role}-emphasis)`,
-              color: `var(--fgColor-onEmphasis)`,
-              borderColor: 'transparent',
-            }}
-          />
-        ))}
-      </Box>
-    </Box>
-  )
-}
